+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.3.0] - 2020-09-03
+
+### Added
+
+- Added better lock support using flock for `LocalEngine`, this can also be disabled through config.
+
+### Changed
+
+- Changed directory mask when creating from 744 to 755
+
 ## [1.2.1] - 2020-07-23
 
 ### Fixed